+++ comandos_para_testes.tmp
@@ -56,3 +56,8 @@ docker run --name dev-postgres \
   -e POSTGRES_DB=devdb \
   -p 5432:5432 \
   -d postgres:latest
+
+
+
+#execução de testes
+go test -v utils_test.go validators.go
