+++ utils/validators.go
@@ -1,9 +1,13 @@
 package utils
 
 import (
+	"log"
 	"strconv"
 	"strings"
 	"time"
+
+	"github.com/Gileno29/clientes-API/models"
+	"gorm.io/gorm"
 )
 
 var StartTime time.Time
@@ -108,4 +112,23 @@ func ValidarCNPJ(cnpj string) bool {
 
 	// Verifica se os dígitos calculados são iguais aos informados
 	return cnpj[12:14] == strconv.Itoa(primeiroDigito)+strconv.Itoa(segundoDigito)
+}
+
+func VerificarTabelaClientes(db *gorm.DB) error {
+
+	if !db.Migrator().HasTable(&models.Cliente{}) {
+		log.Println("Tabela de clientes não encontrada. Criando tabela...")
+
+		err := db.AutoMigrate(&models.Cliente{})
+		if err != nil {
+			log.Fatalf("Erro ao criar tabela de clientes: %v", err)
+			return err
+		}
+
+		log.Println("Tabela de clientes criada com sucesso!")
+	} else {
+		log.Println("Tabela de clientes já existe.")
+	}
+
+	return nil
 }
